requirements.txt
africastalking==1.2.8
asgiref==3.8.1
astroid==3.2.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
dill==0.3.8
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flake8==7.1.1
gunicorn==23.0.0
idna==3.10
isort==5.13.2
josepy==1.14.0
mccabe==0.7.0
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
psycopg2==2.9.9
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pylint==3.2.7
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyOpenSSL==24.2.1
python-dotenv==1.0.1
schema==0.7.7
sqlparse==0.5.1
tomli==2.0.1
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
